Framing out the data model
Before building the database API, it's helpful to think out what the
major data items Boson will be handling are.  This code is by no
means set in stone; it's more meant to think out what the data is,
and as a starting point for thinking about what operations will need
to be performed.  Note that not every data item would be stored in
the database, either; in particular, Request objects would only be
found in the REST API.

This commit also includes a data-model.png graphic (and its associated
source DOT file) that gives a visual picture of what the code
implements in terms of the connections between data objects.