Updates to data model
Determined a solution to the problem of how to look up the relevant,
matching quotas.  The first piece is to reduce the problem space, by
telling Boson which sets of fields are meaningful for looking up
quotas and usages.  The second piece is to store associated authN
and authZ data in the quota and usage as a string, which requires
careful, repeatable serialization and the ability to properly
deserialize the data.