Build the actual models that will be used
The model data is part of the database API; this begins the process
of designing and defining what the database API for Boson is.  Note
that the data_model directory is not yet removed; some items
described there don't correspond to database entities, and the
example is still valuable.  Later commits will remove that code.