Define the database API
The database API is defined as an abstract class, which is extended
by the underlying database drivers to implement the required methods.
We use Python's abc module to ensure that all methods are
implemented, and all the abstract methods are carefully documented so
implementers know the required interface.

Still need to define session and transaction handling, as well as
creating an appropriate factory function which will load the
appropriate database implementation and return it.  Also need to
implement a "simple" database backend, which can be used for testing
purposes.