Add hints arguments and a hints parser
All database query calls now have a "hints" parameter added which
takes a list of the fields the caller wishes to access.  In addition,
a helper hints_parser() method is added, which can parse the list of
fields and returns a dictionary tree, which can be used by the
backend driver to determine what other objects need to be loaded
from the database.